Внес изменения в структуру БД, пересоздал модели
