+++ models/User.php
@@ -2,103 +2,143 @@
 
 namespace app\models;
 
-class User extends \yii\base\BaseObject implements \yii\web\IdentityInterface
+use Yii;
+
+/**
+ * This is the model class for table "users".
+ *
+ * @property int $id
+ * @property string $name
+ * @property string $email
+ * @property string $password
+ * @property string|null $birthday
+ * @property string|null $phone
+ * @property string|null $telegram
+ * @property string|null $information
+ * @property string|null $specializations
+ * @property string|null $avatar
+ * @property string|null $register_date
+ * @property string $role
+ * @property int|null $succesful_tasks
+ * @property int|null $failed_tasks
+ * @property int|null $city_id
+ *
+ * @property Cities $city
+ * @property Responses[] $responses
+ * @property Reviews[] $reviews
+ * @property Reviews[] $reviews0
+ * @property Tasks[] $tasks
+ * @property Tasks[] $tasks0
+ */
+class User extends \yii\db\ActiveRecord
 {
-    public $id;
-    public $username;
-    public $password;
-    public $authKey;
-    public $accessToken;
-
-    private static $users = [
-        '100' => [
-            'id' => '100',
-            'username' => 'admin',
-            'password' => 'admin',
-            'authKey' => 'test100key',
-            'accessToken' => '100-token',
-        ],
-        '101' => [
-            'id' => '101',
-            'username' => 'demo',
-            'password' => 'demo',
-            'authKey' => 'test101key',
-            'accessToken' => '101-token',
-        ],
-    ];
-
-
     /**
      * {@inheritdoc}
      */
-    public static function findIdentity($id)
+    public static function tableName()
     {
-        return isset(self::$users[$id]) ? new static(self::$users[$id]) : null;
+        return 'users';
     }
 
     /**
      * {@inheritdoc}
      */
-    public static function findIdentityByAccessToken($token, $type = null)
+    public function rules()
     {
-        foreach (self::$users as $user) {
-            if ($user['accessToken'] === $token) {
-                return new static($user);
-            }
-        }
-
-        return null;
+        return [
+            [['name', 'email', 'password', 'role'], 'required'],
+            [['birthday', 'register_date'], 'safe'],
+            [['information', 'role'], 'string'],
+            [['succesful_tasks', 'failed_tasks', 'city_id'], 'integer'],
+            [['name'], 'string', 'max' => 150],
+            [['email', 'password', 'phone', 'telegram'], 'string', 'max' => 100],
+            [['specializations', 'avatar'], 'string', 'max' => 255],
+            [['email'], 'unique'],
+            [['city_id'], 'exist', 'skipOnError' => true, 'targetClass' => Cities::class, 'targetAttribute' => ['city_id' => 'id']],
+        ];
     }
 
     /**
-     * Finds user by username
+     * {@inheritdoc}
+     */
+    public function attributeLabels()
+    {
+        return [
+            'id' => 'ID',
+            'name' => 'Name',
+            'email' => 'Email',
+            'password' => 'Password',
+            'birthday' => 'Birthday',
+            'phone' => 'Phone',
+            'telegram' => 'Telegram',
+            'information' => 'Information',
+            'specializations' => 'Specializations',
+            'avatar' => 'Avatar',
+            'register_date' => 'Register Date',
+            'role' => 'Role',
+            'succesful_tasks' => 'Succesful Tasks',
+            'failed_tasks' => 'Failed Tasks',
+            'city_id' => 'City ID',
+        ];
+    }
+
+    /**
+     * Gets query for [[City]].
      *
-     * @param string $username
-     * @return static|null
+     * @return \yii\db\ActiveQuery
      */
-    public static function findByUsername($username)
+    public function getCity()
     {
-        foreach (self::$users as $user) {
-            if (strcasecmp($user['username'], $username) === 0) {
-                return new static($user);
-            }
-        }
-
-        return null;
+        return $this->hasOne(Cities::class, ['id' => 'city_id']);
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAuthKey()
-    {
-        return $this->authKey;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function validateAuthKey($authKey)
-    {
-        return $this->authKey === $authKey;
-    }
-
-    /**
-     * Validates password
+     * Gets query for [[Responses]].
      *
-     * @param string $password password to validate
-     * @return bool if password provided is valid for current user
+     * @return \yii\db\ActiveQuery
      */
-    public function validatePassword($password)
+    public function getResponses()
     {
-        return $this->password === $password;
+        return $this->hasMany(Responses::class, ['executor_id' => 'id']);
+    }
+
+    /**
+     * Gets query for [[Reviews]].
+     *
+     * @return \yii\db\ActiveQuery
+     */
+    public function getReviews()
+    {
+        return $this->hasMany(Reviews::class, ['customer_id' => 'id']);
+    }
+
+    /**
+     * Gets query for [[Reviews0]].
+     *
+     * @return \yii\db\ActiveQuery
+     */
+    public function getReviews0()
+    {
+        return $this->hasMany(Reviews::class, ['executor_id' => 'id']);
+    }
+
+    /**
+     * Gets query for [[Tasks]].
+     *
+     * @return \yii\db\ActiveQuery
+     */
+    public function getTasks()
+    {
+        return $this->hasMany(Tasks::class, ['customer_id' => 'id']);
+    }
+
+    /**
+     * Gets query for [[Tasks0]].
+     *
+     * @return \yii\db\ActiveQuery
+     */
+    public function getTasks0()
+    {
+        return $this->hasMany(Tasks::class, ['executor_id' => 'id']);
     }
 }
