--- models/Users.php
@@ -1,145 +1,0 @@
-<?php
-
-namespace app\models;
-
-use Yii;
-
-/**
- * This is the model class for table "users".
- *
- * @property int $id
- * @property string $name
- * @property string $email
- * @property string $password
- * @property string|null $birthday
- * @property string|null $phone
- * @property string|null $telegram
- * @property string|null $information
- * @property string|null $specializations
- * @property string|null $avatar
- * @property string|null $register_date
- * @property string $role
- * @property int|null $succesful_tasks
- * @property int|null $failed_tasks
- * @property int|null $city_id
- *
- * @property Cities $city
- * @property Responses[] $responses
- * @property Reviews[] $reviews
- * @property Reviews[] $reviews0
- * @property Tasks[] $tasks
- * @property Tasks[] $tasks0
- */
-class Users extends \yii\db\ActiveRecord
-{
-    /**
-     * {@inheritdoc}
-     */
-    public static function tableName()
-    {
-        return 'users';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function rules()
-    {
-        return [
-            [['name', 'email', 'password', 'role'], 'required'],
-            [['birthday', 'register_date'], 'safe'],
-            [['role'], 'string'],
-            [['succesful_tasks', 'failed_tasks', 'city_id'], 'integer'],
-            [['name'], 'string', 'max' => 150],
-            [['email', 'password', 'phone', 'telegram'], 'string', 'max' => 100],
-            [['information'], 'string', 'max' => 1],
-            [['specializations', 'avatar'], 'string', 'max' => 255],
-            [['email'], 'unique'],
-            [['city_id'], 'exist', 'skipOnError' => true, 'targetClass' => Cities::class, 'targetAttribute' => ['city_id' => 'id']],
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => 'ID',
-            'name' => 'Name',
-            'email' => 'Email',
-            'password' => 'Password',
-            'birthday' => 'Birthday',
-            'phone' => 'Phone',
-            'telegram' => 'Telegram',
-            'information' => 'Information',
-            'specializations' => 'Specializations',
-            'avatar' => 'Avatar',
-            'register_date' => 'Register Date',
-            'role' => 'Role',
-            'succesful_tasks' => 'Succesful Tasks',
-            'failed_tasks' => 'Failed Tasks',
-            'city_id' => 'City ID',
-        ];
-    }
-
-    /**
-     * Gets query for [[City]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getCity()
-    {
-        return $this->hasOne(Cities::class, ['id' => 'city_id']);
-    }
-
-    /**
-     * Gets query for [[Responses]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getResponses()
-    {
-        return $this->hasMany(Responses::class, ['executor_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[Reviews]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getReviews()
-    {
-        return $this->hasMany(Reviews::class, ['customer_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[Reviews0]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getReviews0()
-    {
-        return $this->hasMany(Reviews::class, ['executor_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[Tasks]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getTasks()
-    {
-        return $this->hasMany(Tasks::class, ['customer_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[Tasks0]].
-     *
-     * @return \yii\db\ActiveQuery
-     */
-    public function getTasks0()
-    {
-        return $this->hasMany(Tasks::class, ['executor_id' => 'id']);
-    }
-}
