+++ shema.sql
@@ -21,7 +21,7 @@ CREATE TABLE users (
     birthday DATETIME,
     phone VARCHAR(100),
     telegram VARCHAR(100),
-    information CHAR,
+    information TEXT,
     specializations VARCHAR(255),
     avatar VARCHAR(255),
     register_date DATETIME DEFAULT CURRENT_TIMESTAMP,
@@ -44,11 +44,9 @@ CREATE TABLE tasks (
     id INT AUTO_INCREMENT PRIMARY KEY,
     customer_id INT NOT NULL,
     title VARCHAR(255) NOT NULL,
-    description CHAR,
+    description TEXT,
     category_id INT NOT NULL,
     city_id INT,
-    latitude DECIMAL(9, 7),
-    longtitude DECIMAL(9, 7),
     budget VARCHAR(100),
     deadline DATETIME,
     creation_date DATETIME DEFAULT CURRENT_TIMESTAMP,
@@ -73,7 +71,7 @@ CREATE TABLE responses (
     id INT AUTO_INCREMENT PRIMARY KEY,
     executor_id INT NOT NULL,
     task_id INT NOT NULL,
-    comment CHAR,
+    comment TEXT,
     price INT,
     FOREIGN KEY (executor_id) REFERENCES users(id),
     FOREIGN KEY (task_id) REFERENCES tasks(id)
@@ -85,7 +83,7 @@ CREATE TABLE reviews (
     customer_id INT NOT NULL,
     executor_id INT NOT NULL,
     task_id INT NOT NULL,
-    comment CHAR,
+    comment TEXT,
     grade INT,
     creation_date DATETIME DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (customer_id) REFERENCES users(id),
